firebase with google,fb and email
version 0.0.0.0.3
